+++ src/core/config/authorization.js
@@ -1,0 +1,5 @@
+import { AuthorizationService } from 'packages/authModel/module/authorization/AuthorizationService';
+import { Rules } from '../rules';
+
+export const authorization = AuthorizationService
+                                .builder(Rules);