+++ src/core/index.js
@@ -4,6 +4,8 @@ import express from 'express';
 import { HttpExceptionFilter } from 'packages/httpException/HttpExceptionFilter';
 import { SecurityFilter } from 'packages/authModel/core/security/SecurityFilter';
 import { InvalidUrlFilter } from 'packages/handler/filter/InvalidUrlFilter';
+import { AuthorizationLookup } from 'packages/authModel/module/authorization/AuthorizationLookup';
+import { AuthorizationValidator } from 'packages/authModel/module/authorization/AuthorizationValidator';
 import { ApiDocument } from './config/swagger';
 import { AppBundle } from './config';
 import { ModuleResolver } from './api';
@@ -19,6 +21,9 @@ const app = express();
         .applySwagger(ApiDocument)
         .applyGlobalFilters([new HttpExceptionFilter(), new InvalidUrlFilter()])
         .run();
+
+    const container = await AuthorizationLookup.builder().collect();
+    AuthorizationValidator.addAuthorizeStore(container);
 })();
 
 export default app;