+++ src/packages/authModel/common/utils/MapperUtil.js
@@ -1,0 +1,5 @@
+export class MapperUtil {
+    static map(accessors, store) {
+        return accessors.map(accessor => store[accessor]);
+    }
+}